hotorcold/requirements.txt
Django==1.11.4
djangorestframework==3.6.3
psycopg2==2.7.3
pytest-django==3.1.2
celery==4.1.0